add explicit rescale button for ylimits and colorscale
