+++ python/src/scipp/plot/tools.py
@@ -71,27 +71,27 @@ def parse_params(params=None,
         for key, val in params.items():
             parsed[key] = val
 
-    need_norm = False
-    # TODO: sc.min/max currently return nan if the first value in the
-    # variable array is a nan. Until sc.nanmin/nanmax are implemented, we fall
-    # back to using numpy, both when a Variable and a numpy array are supplied.
-    if variable is not None:
-        _find_min_max(variable.values, parsed)
-        need_norm = True
-    if array is not None:
-        _find_min_max(array, parsed)
-        need_norm = True
+    # need_norm = False
+    # # TODO: sc.min/max currently return nan if the first value in the
+    # # variable array is a nan. Until sc.nanmin/nanmax are implemented, we fall
+    # # back to using numpy, both when a Variable and a numpy array are supplied.
+    # if variable is not None:
+    #     _find_min_max(variable.values, parsed)
+    #     need_norm = True
+    # if array is not None:
+    #     _find_min_max(array, parsed)
+    #     need_norm = True
 
-    if need_norm:
-        if min_val is not None:
-            parsed["vmin"] = min(parsed["vmin"], min_val)
-        if max_val is not None:
-            parsed["vmax"] = max(parsed["vmax"], max_val)
-        if parsed["log"]:
-            norm = LogNorm
-        else:
-            norm = Normalize
-        parsed["norm"] = norm(vmin=parsed["vmin"], vmax=parsed["vmax"])
+    # if (variable is not None) or (array is not None):
+    #     if min_val is not None:
+    #         parsed["vmin"] = min(parsed["vmin"], min_val)
+    #     if max_val is not None:
+    #         parsed["vmax"] = max(parsed["vmax"], max_val)
+    if parsed["log"]:
+        norm = LogNorm
+    else:
+        norm = Normalize
+    parsed["norm"] = norm(vmin=parsed["vmin"], vmax=parsed["vmax"])
 
     # Convert color into custom colormap
     if parsed["color"] is not None:
@@ -108,21 +108,21 @@ def make_fake_coord(dim, size, unit=None):
     return sc.Variable(dims=[dim], **args)
 
 
-def _find_min_max(array, params):
-    if params["vmin"] is None or params["vmax"] is None:
-        if params["log"]:
-            with np.errstate(divide="ignore", invalid="ignore"):
-                valid = np.ma.log10(array)
-        else:
-            valid = np.ma.masked_invalid(array, copy=False)
-    if params["vmin"] is None:
-        params["vmin"] = valid.min()
-        if params["log"]:
-            params["vmin"] = 10.0**params["vmin"]
-    if params["vmax"] is None:
-        params["vmax"] = valid.max()
-        if params["log"]:
-            params["vmax"] = 10.0**params["vmax"]
+# def _find_min_max(array, params):
+#     if params["vmin"] is None or params["vmax"] is None:
+#         if params["log"]:
+#             with np.errstate(divide="ignore", invalid="ignore"):
+#                 valid = np.ma.log10(array)
+#         else:
+#             valid = np.ma.masked_invalid(array, copy=False)
+#     if params["vmin"] is None:
+#         params["vmin"] = valid.min()
+#         if params["log"]:
+#             params["vmin"] = 10.0**params["vmin"]
+#     if params["vmax"] is None:
+#         params["vmax"] = valid.max()
+#         if params["log"]:
+#             params["vmax"] = 10.0**params["vmax"]
 
 
 
@@ -133,27 +133,27 @@ def vars_to_err(v):
     v[non_finites] = 0.0
     return v
 
-def get_finite_y(arr, logy=False):
-    if logy:
-        with np.errstate(divide="ignore", invalid="ignore"):
-            arr = np.log10(arr, out=arr)
-    subset = np.where(np.isfinite(arr))
-    return arr[subset]
+# def get_finite_y(arr, logy=False):
+#     if logy:
+#         with np.errstate(divide="ignore", invalid="ignore"):
+#             arr = np.log10(arr, out=arr)
+#     subset = np.where(np.isfinite(arr))
+#     return arr[subset]
 
-def get_ylim(var=None, ymin=None, ymax=None, errorbars=False, logy=False):
-    if errorbars:
-        err = vars_to_err(var.variances)
-    else:
-        err = 0.0
+# def get_ylim(var=None, ymin=None, ymax=None, errorbars=False, logy=False):
+#     if errorbars:
+#         err = vars_to_err(var.variances)
+#     else:
+#         err = 0.0
 
-    ymin_new = np.amin(get_finite_y(var.values - err, logy=logy))
-    ymax_new = np.amax(get_finite_y(var.values + err, logy=logy))
+#     ymin_new = np.amin(get_finite_y(var.values - err, logy=logy))
+#     ymax_new = np.amax(get_finite_y(var.values + err, logy=logy))
 
-    dy = 0.05 * (ymax_new - ymin_new)
-    ymin_new -= dy
-    ymax_new += dy
-    if logy:
-        ymin_new = 10.0**ymin_new
-        ymax_new = 10.0**ymax_new
-    return [ymin_new if ymin is None else min(ymin, ymin_new),
-            ymax_new if ymax is None else max(ymax, ymax_new)]
+#     dy = 0.05 * (ymax_new - ymin_new)
+#     ymin_new -= dy
+#     ymax_new += dy
+#     if logy:
+#         ymin_new = 10.0**ymin_new
+#         ymax_new = 10.0**ymax_new
+#     return [ymin_new if ymin is None else min(ymin, ymin_new),
+#             ymax_new if ymax is None else max(ymax, ymax_new)]
